backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-parler==2.3
django-parler-rest==2.2
django-tinymce==3.5.0
djangorestframework==3.13.1
idna==3.3
pymongo==4.2.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
django-cors-headers==3.13.0
docx2pdf==0.1.8
docxtpl==0.16.4
pillow==9.0.0
lxml==4.9.1
requests==2.28.1
beautifulsoup4==4.11.1
django-celery==3.3.1